requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-pptx
openai
langchain
faiss-cpu
python-dotenv
langchain-community
tiktoken